Patch V2
I think it's better to recover the real name of the executable that we download.

It will be better for multiple download on the near future, instead of to rename
the file downloaded with "{}.exe".format(self.game.name)

If we keep this (old) solution for multiple download, the setup file downloaded
at 1st will be named "Anno 1404.exe" and the .bin file will have the same name
and same extension.

With my solution, the setup and .bin files for Anno 1404 (and others games)
will be named with their original name :
- setup_anno_1404_2.01_v2_(30326).exe
- setup_anno_1404_2.01_v2_(30326)-1.bin