Moved to separate methods and use absolute paths.
Made a function and a method, maybe the function could even be moved to a more generic .py file (but for clarity, left it in the same file for now). The method is specific for the downloads, also included a buffer in case a barely sufficient disk space still leads to users' OS slowing down. Also included real paths, so users using symlinks don't get false error messages.

Closing the app, making space, opening the app seems to be enough to let downloads continue as before after raising an error.