Don't try updating when game version in API is set to null
For certain games the version in the API is literally set to null, which
will be read in python as None. When this happens the game will always
show that there is an update. This commit does a check so see if the
version in the api is set before returning that there is an update.