Fix freeze for games generating a lot of output
When using minigalaxy, some games tend to freeze after a few minutes.
When running them manually, from a terminal, I can't reproduce the
freeze. This is the case, at least, for "Darkest Dungeon" and "Slay the
Spire".
This issue seems to be related to how the `subprocess.Popen` call is
done when launching the game. `stdout` and `stderr` are set to
`subprocess.PIPE`, but `subprocess.PIPE.communicate` is called only once
if the game is not running after a 3s delay. Child output is filling an
internal buffer and the game hangs when this buffer is full.

Now, stderr is merged in stdout, and child stdout is consummed and sent
to parent stdout.