Reduce occurrence of download/install race condition
If an installer deletes the directory a new download has already run
prepare_location for while it is starting the download_request stream the open
method will fail. By creating the file in prepare_location and deleting the
directory after an install only if the directory is empty, the times the race
condition occurs is reduced.