Use threads to speed up connectivity check
Uses a ThreadPoolExecutor to send the requests to embed.gog.com and auth.gog.com from separate threads. As these threads do I/O-bound work, they will hand the GIL back while waiting for responses, which speeds up the work as opposed to sending the requests and waiting for the responses sequentially.