Adding Flatpak to README
I considered making this an issue, but I figured I could just write the change, and you can decline the pull request if you don't want it. 

I only noticed today that Minigalaxy has what seems to be an official & maintained release on Flathub. However, it is not mentioned on the README/website. I added a section under Installation for Flatpak like the other entries. This would help make sure people don't miss this option.