provide utility for easier maintenance of ongoing downloads in Config
Config.current_downloads is a json list, but effectively used like a
set. But there is no centralized management to handle this. Instead,
multiple different implementations to add or remove entries are littered
across LibraryEntry and GameTile(View):
 * `__download`
 * `prevent_resume_on_startup`
 * `get_download_info`

These implementations all iterate over the list and build a completely
new one while omitting an item.
One implementation first loads the active list as a set before
re-assigning to the property as list. This has the potential side effect
that the order will be changed because set order is unspecified.
Another one iterates over the list to search for an entry instead of
just using `if value in list:`

`__install_game` gets the list and removes an entry, but this is not
reflected in the config.json until any other property setter triggers a
save.