minor improvements in wine install
1. Speed up initial prefix creation:
When a new prefix is created during install, it already is fully
updated.
The additional command 'wineboot -u' forced running the same upgrade
procedure a second time which can take a long time, depending on the
hardware.
Since the main point of this command was just to run anything in wine to
let it create the prefix, it can be changed to something that works
faster.
2. In addition to using an env variable to disable winemenubuilder, it
is tried to disable it more permanently in the registry to prevent
additional installations triggered by the game itself from adding more
shortcuts (some game engines install dependencies they need afterwards).
Combined with 1., so the first action that is done to create the
wineprefix is running a regedit import