fix some bugs with the new keep_installers dir structure
verify_installer_integrity also requires the file list of an installer,
therefore it is now mandatory and will be generated by install_game from
the installer name if not given.

LibraryEntry.get_keep_executable_path must filter out directory names
because the check if a dir entry has executable flag set OR an ending is
not sufficient. Directories need to be executable to be opened.

Use os.path.basename to get last directory in installer path to prevent
potential issues with trailing slashes

make sure named argument 'to_path' is used consequently when stripping a
game's name of problematic characters