several bug fixes with queued installs
1. cache and keep_installer: only delete files when install was
successful or failed because of checksums.
  Stuff like not enough disk space should not delete files which are
otherwise valid, that's just a waste of data plan and time. Even if the
files are in cache and should not be kept permanently.

2. keep_installers + invalid checksums:
  delete invalid files and mark the game to force redownload it.
  Otherwise, only having + deleting a broken .bin will not enable the
download button again, because there'd still be an executable in
keep_path

3. install_game was reworked and has an optional way to exit on error by
raising an exception instead.
  This way, several different kind of installation errors can be
reported type-safe, which actually is required to properly treat the
checksum exception.
  The current solution uses a wrapper function, but it would be better
to fully change the implementation of all install steps to raise instead
of returning error messages.
  Only used by installations coming from the install queue.

4. shifted order of some install steps around. Creating a desktop file
now is the last step before installer file deletion. And, with the
exception of checksum errors, installer files will be deleted only on
success.