Fix STOPPED downloads can not be 'deleted'
When a download that is still in the queue gets stopped, there won't be
a partial download.
So checking for an existing partial file to allow going from STOPPED ->
CANCELED would not work and block this download from progressing in
state.
Also, the cancel state must be set for the download even if it is not
active